fix(gl): CI gate fixes for themed cursors (#178)
- Split the RENDER cursor wiring out of platform_x11.go into
  cursors_x11.go (large-files gate: >800 lines).
- Mark the three theme-file reads with #nosec G304 (fixed subdirs
  under the user's own config; gosec gate).
- Fix the XSETTINGS color-entry builder in the tests, which emitted a
  malformed blob and made TestParseXSettings depend on map iteration
  order (caught by the Coverage job).
